project/requirements.txt
django == 1.9.6
djangorestframework
MySQL-python